🚚 Move google-identity-platform functions to a subdirectory
