+++ src/functions/google-identity-platform/index.ts
@@ -1,0 +1,2 @@
+export { GoogleIdentityPlatformGenerateCustomToken } from './generate-custom-token.js';
+export { GoogleIdentityPlatformGenerateToken } from './generate-token.js';
